refactor: enum 제거
