+++ src/main/java/com/techeer/fashioncloud/domain/geo/controller/AddressController.java
@@ -8,6 +8,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -21,6 +22,7 @@ public class AddressController {
     private final AddressService addressService;
 
     @GetMapping()
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     @Operation(summary = "위치 반환", description ="위경도를 이용해 위치를 반환한다.")
     public ResponseEntity<ResultResponse> getAddressHere(
             @Parameter(name="latitude") @RequestParam Double latitude,
